Fix for CardLibraryScreen moving in week 31 patch
